backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
dj-database-url
gunicorn
psycopg[binary]
whitenoise
sqlparse==0.5.5
